// src/global/enseignant.d.ts
export interface Enseignant{
    
    noEnseignant : number;
    adresse : String ;
	codePostal : String ;
	emailPerso : String ;
	emailUbo : String ;
	mobile : String ;
	nom : String ;
	pays : String ;
	prenom : String ;
	sexe : String ;
	telephone : String ;
	type : String ;
    ville : String ;

}